requirements.txt
fastapi
uvicorn[standard]
python-dotenv==1.0.0
pandas==2.1.3
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
slowapi==0.1.8
aiohttp==3.12.12
motor
google-cloud-pubsub
jose
python-jose
google-cloud-storage


